Refactoring the app to use Core Data.

+++ Vesta/Model/Persistence.swift
@@ -17,7 +17,7 @@ struct PersistenceController {
             let newTask = Task(context: viewContext)
             newTask.id = UUID()
             newTask.title = "Task #\(index + 1)"
-            newTask.userOrder = Int16(index + 1)
+            newTask.userOrder = Int64(index + 1)
         }
         do {
             try viewContext.save()
