+++ Vesta/Model/Task+CoreDataProperties.swift
@@ -18,7 +18,7 @@ extension Task {
 
     @NSManaged public var id: UUID?
     @NSManaged public var title: String?
-    @NSManaged public var userOrder: Int16
+    @NSManaged public var userOrder: Int64
     @NSManaged public var timeEntries: NSSet?
 
 }
